+++ main.py
@@ -2,7 +2,7 @@ import sys
 from PyQt6.QtWidgets import QApplication
 
 from classes.logger import Logger
-from classes.main_window import MainWindow
+from gui.main_window import MainWindow
 from classes.trigger_bot import CS2TriggerBot
 
 def main():